tutorials/ContextEngineering/requirements.txt
langchain==0.2.11
langchain_aws==0.1.12
langchain_core==0.2.23
python-dotenv==1.0.1
redis==5.0.7
streamlit==1.33.0
